+++ backend/tests/test_smoke_azure_ai_document_intelligence.py
@@ -2,7 +2,7 @@ from logging import getLogger
 
 import pytest
 
-from tests.utilities import RUN_TEST, client, image
+from backend.tests.utilities import RUN_TEST, client, image
 
 logger = getLogger(__name__)
